enum table deleted, deleted dev folder,moved data into other sample data files
