+++ quickboosters/api/users/dev/sample_data.py
@@ -16,11 +16,11 @@ def create_user():
     except Exception as e:
         print(str(e))
 
-def create_order():
-  order = Orders(order_type="solo_boost",user_id='1',order_amount=69.42,status='completed',date_ordered=datetime.datetime.now(),game='league of legends')
-  print(order.user_id)
-  db.session.add(order)
-  db.session.commit()
+#def create_order():
+  #order = Orders(order_type="solo_boost",user_id='1',order_amount=69.42,status='completed',date_ordered=datetime.datetime.now(),game='league of legends')
+  #print(order.user_id)
+  #db.session.add(order)
+  #db.session.commit()
 
 def Chatlog():
     Chatlog = ChatLog(message='message',userfrom='user.username',created_date=datetime.datetime.now())
@@ -28,11 +28,11 @@ def Chatlog():
     db.session.add(Chatlog)
     db.session.commit()
 
-def order_status():
-    status = Order_Status('status': order_status.enum, 'order_id': order_status.id, 'game': order_status.game)
-    print(status)
-    db.session.add(status)
-    db.session.commit()
+#def order_status():
+   # status = Order_Status('status': order_status.enum, 'order_id': order_status.id, 'game': order_status.game)
+   # print(status)
+   # db.session.add(status)
+   # db.session.commit()
